==> notebook/3.1 - DATA PREPARATION (last_fm dataset).ipynb <==
# DATA PREPARATION LAST FM

FICHIER: *'userid-timestamp-artid-artname-traid-traname.tsv'*

## Imports 

import pandas as pd

## Data import

file = 'raw data/lastfm-dataset-1K/userid-timestamp-artid-artname-traid-traname.tsv'
df =  pd.read_csv(file, sep='\t',error_bad_lines=False, header = None)

# on renome les colonnes
df.rename(columns={0:'user_id', 1:'time_stamp', 2:'artist_id', 3:'artist_name', 4:'track_id', 5:'track_name'}, inplace=True)

df.head()

## Quick Exploration 

print('shape:',df.shape)
print('user:',df['user_id'].nunique())
print('trak:',df['track_name'].nunique())

# les 10 tack_name les plus écoutés
df['track_name'].value_counts()[0:10]

# les 10 artis_name les plus écoutés
df['artist_name'].value_counts()[0:10]

# Date de première entrée et de dernière entrée
print('Start: ', df['time_stamp'].tail(1).values[0])
print('Stop:  ',df['time_stamp'].head(1).values[0])

### Reset des IDs (verst des int)

# On ne garde que la partie numérique des id
df['user_id'] = df['user_id'].apply(lambda x: int(x[5:]))

# reset des id par des int
d=dict()
for i,name in enumerate(df['artist_id'].unique()):
    d[name]=i
    
df['artist_id'] = df['artist_id'].apply(lambda x: d[x])

# creation d'un nouveau track_id (car trop de NAN)

#on crés une serie uneique de artist+track pour eviter les erreus sur les tracks de même nom
a = df['artist_name']+df['track_name'] 

d=dict()
for i,name in enumerate(a.unique()):
    d[name]=i
    
df['track_id'] = a.apply(lambda x: d[x])

### Verification du reset

df.head()

df[df['track_id']==6425].head()

## Filtre des tracks

df['track_id'].value_counts().describe()

df_temp = df.groupby(['track_id','user_id']).size().reset_index()
df_temp = df_temp.groupby(['track_id']).size().reset_index()
df_temp.rename(columns={0:'count'}, inplace=True)

track_to_keep = set(df_temp[df_temp['count']>=10]['track_id'].values) # 10 écoutes

print('Nb de track à garder: ', len(track_to_keep))

filter = df['track_id'].apply(lambda x: x in track_to_keep)

df = df[filter].copy()

print('shape:',df.shape)
print('user:',df['user_id'].nunique())
print('trak:',df['track_id'].nunique())

## Filtre des utilisateurs 

df['user_id'].value_counts().describe()

df_temp = df.groupby(['track_id','user_id']).size().reset_index()
df_temp = df_temp.groupby(['user_id']).size().reset_index()
df_temp.rename(columns={0:'count'}, inplace=True)

user_to_keep = set(df_temp[df_temp['count']>50]['user_id'].values)

print('Nb utilisateur à garder:' , len(user_to_keep))

filter = df['user_id'].apply(lambda x: x in user_to_keep)

df = df[filter].copy()

print('shape:',df.shape)
print('user:',df['user_id'].nunique())
print('trak:',df['track_id'].nunique())

## Reset des IDs track 
Pour optimiser la taille des matrices des model de colaborative filtering

# reset des id par des int
d=dict()
for i,name in enumerate(df['track_id'].unique()):
    d[name]=i
    
df['track_id'] = df['track_id'].apply(lambda x: d[x])

df['track_id'].nunique()

# Sauvegarde du Nouveau dataset

df.shape

df.to_csv('filtered data\\filtered_fm_dataset.csv')
