==> notebook/3.2 - DATA PREPARATION (Spotify API).ipynb <==
# DATA PREPARATION LAST FM (Spotify API)

### Imports 

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from time import sleep
import json

import pandas as pd

### Class permetant de scraper l'API Spotify 

class SpotifyScraper():
    
    '''
    Class permetant de récuperer les features descriptives d'un track à partir du titre et du l'artiste
    
    Pour fonctionner un compte Spotify for Developpers est necessaire
    '''
    
    def __init__(self, cid, secret):
        
        '''
        Initialisation de la class
        
        cid --> identifiant du compre Spotify for Developpers
        secret --> clée secrete generée pour l'application
        
        cf Spotify for Developpers
        '''
        
        client_credentials_manager = SpotifyClientCredentials(client_id=cid, client_secret=secret)
        self.spotify = spotipy.Spotify(client_credentials_manager = client_credentials_manager, requests_timeout=10)
    
    def get_features(self, song_id):
        
        '''
        Récupére les feartures d'un titre à partir de son id Spotify 
        exemple id : '5ghIJDpPoe3CfHMGu71E6T'
        
        retourne un liste de features (float)
        '''
        
        sleep(0.1) # Tenmps d'attente pour eviter de saturer l'API
        
        try:
            features = ['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
                        'acousticness', 'instrumentalness', 'liveness',
                        'valence', 'tempo']

            track = self.spotify.audio_features([song_id])
            result = list()

            for feature in features:
                a = track[0][feature]
                result.append(a)

            return result
    
            
        except:
            return None
    
    def search_song(self, txt, artist=None):
        
        '''
        recherhce un titre à partirt du titre de la chanson et d'un nom d'artise
        retourne l'id track Spotify correspondant si il est trouvé
        '''
        
        try:
            ans = self.spotify.search(q='track:' + txt, type='track')
            
            if artist==None:
                return ans['tracks']['items'][0]['id']

            for tr in ans['tracks']['items']:
                
                if tr['name'].lower() in txt.lower():
                    for art in tr['artists']:
                        
                        if artist.lower() in art['name'].lower():
                            return tr['id'], art['name'], tr['name']
            
            return None
        
        except:
            return None
        
    
    def search_song_features(self, txt, artist=None):
        
        '''
        recherhce un titre à partirt du titre de la chanson et d'un nom d'artise
        puis récupère les features associées
        retourne un dictionaire
        '''
        
        sleep(0.1) # Tenmps d'attente pour eviter de saturer l'API
        
        try:
            
            id, name, artist = self.search_song(txt, artist)
            features = self.get_features(id)
        
            return {'artist':artist, 'name': name, 'features': features}
        
        except:
            return None
        

### Creation d'un scraper

with open ('credential/Spotify_credential.json') as f:
    credentials = json.load(f)

# creation d'un scraper
scraper = SpotifyScraper(credentials['cid'], credentials['secret'])

# test de la méthode get_features
scraper.get_features('5ghIJDpPoe3CfHMGu71E6T')

# test de la méthode search_song_features
scraper.search_song_features('The Girl And The Robot', 'Röyksopp')

## Chargement du dataset 

df = pd.read_csv('filtered data/filtered_fm_dataset.csv', index_col=0)
df = df.groupby(['track_id']).first().reset_index()

track_names = df['track_name'].values
artist_names = df['artist_name'].values
track_ids = df['track_id'].values

## Creation de deux fichiers temporaire

Le grand nombre de donnés et le temps de traitement (environ 10h) nous oblige à faire un fichier de dump <br/>
Pour eviter d'être bannie, nous traiton les données en batch de 10k ~ 1h de traitement

**Fichiers:**<br/>
out.csv --> fichier txt où les features sont ajoutés les une aprés les autres <br/>
ou_id.csv --> fichier regroupant les id déjà traités

col = ['artist_name','track_name','track_id','danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
                        'acousticness', 'instrumentalness', 'liveness',
                        'valence', 'tempo']

with open('out.csv', 'w') as f:
                f.write(';'.join(col)+'\n')

try:
    with open('id_out.csv', 'r') as f:
        out_id = set([int (l) for l in f.readlines()])
    print(out_id)
except:
    out_id=set()

from tqdm import tqdm

col = ['artist_name','track_name','track_id','danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
                        'acousticness', 'instrumentalness', 'liveness',
                        'valence', 'tempo']

# recuperation des tack_id déjà traités
try:
    with open('id_out.csv', 'r') as f:
        out_id = set([int (l) for l in f.readlines()])
except:
    out_id=set()


# taille du batch
max=100

for artist, track,id in tqdm(zip(artist_names, track_names, track_ids)) :
    
    if id not in out_id:
        
        if max==0:
            break
        
        result = scraper.search_song_features(track, artist)
        max-=1

        try:
            with open('id_out.csv', 'a+') as f:
                f.write(str(id)+'\n')
            
            line = f'{artist};{track};{id};'+';'.join([str(x) for x in result['features']])+'\n'
            with open('out.csv', 'a+') as f:
                f.write(line)
        except:
                pass
    else:
        pass

## Performance 

with open('out.csv', 'r') as f:
    aa = len(f.readlines())

with open('id_out.csv', 'r') as f:
    bb = len(f.readlines())

print(aa/bb)

## Transformation du fichier dump out.csv en DataFrame

columns = ['artist_name','track_name','track_id','danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
                        'acousticness', 'instrumentalness', 'liveness',
                        'valence', 'tempo']

df_2 = pd.DataFrame(columns=columns)

with open('out.csv', 'r') as f:
    temp = dict()
    for line in f.readlines()[1:]:
        for col, val in zip(columns,line[:-1].split(';')):
            temp[col] = val
        df_2 = df_2.append(pd.DataFrame.from_dict([temp]))
        

df_2.head()

## Rectification des types des variables

columns = ['track_id', 'key', 'mode']

for col in columns:
    df_2[col] = df_2[col].astype('int')

columns = ['danceability', 'energy', 'loudness', 'speechiness',
                        'acousticness', 'instrumentalness', 'liveness',
                        'valence', 'tempo']

for col in columns:
        df_2[col] = df_2[col].apply(lambda x: x.replace(',','.'))

columns = ['danceability', 'energy', 'loudness', 'speechiness',
                        'acousticness', 'instrumentalness', 'liveness',
                        'valence', 'tempo']

for col in columns:
    df_2[col] = df_2[col].astype('float')

df_2 = df_2.reset_index().drop(columns=['index'])

df_2.to_csv('filtered data/track_features_fm_dataset.csv')

## Verification du DataFrame

df_2 = pd.read_csv('filtered data/track_features_fm_dataset.csv', index_col=0)
df_2.head(10)

df_2.shape

df_2.info()
