==> notebook/3.3 - DATA PREPARATION (Rating).ipynb <==
# DATA PREPARATION LAST FM

FICHIER: *'userid-timestamp-artid-artname-traid-traname.tsv'*

## Imports 

import pandas as pd
import numpy as np

## Data import

df = pd.read_csv('filtered data/filtered_fm_dataset.csv', error_bad_lines=False)
df = df.iloc[:,1:]

### Calcul du nomnre d'écoute pour un track_id par utilisateur

df_track = df.groupby(['user_id','track_id']).size().reset_index(name='listening_count')

g = df_track.groupby('user_id')["listening_count"]
min_, max_ = g.transform('min'), g.transform('max')

### Normalisation du nombre d'écoutes 

df_track["listening_count_normalised"] = (df_track["listening_count"] - min_) / (max_ - min_)*100
df_track["listening_count_normalised"] = df_track["listening_count_normalised"].fillna(0.0).astype(int)+1

df_final_jg=df_track[["user_id","track_id","listening_count_normalised"]]
df_final_jg=df_final_jg.rename(columns={"listening_count_normalised":"rating"})

df_rating = pd.merge(
            df,
            df_final_jg,
            how="left",
            left_on=['user_id', 'track_id'],
            right_on=['user_id', 'track_id']
        )

df_rating.head()

df.head()

df_rating.to_csv('filtered data/filtered_rating_fm_dataset.csv',index=False,float_format='%.2f')
