==> notebook/3.4 - DATA PREPARATION (train_test_split).ipynb <==
# DATA PREPARATION Train_test_split

FICHIER: *'filtered_rating_fm_dataset.csv'*

## Imports

import random
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

## Data imports

df = pd.read_csv('filtered data/filtered_rating_fm_dataset.csv')
df.head()

## Creation d'une matrice d'interaction Item/user/weight

df_gb = df.groupby(['user_id','track_id']).mean().reset_index()

item = np.array(df_gb[['track_id']].values).T[0]
user = np.array(df_gb[['user_id']].values).T[0]
weight = np.array(df_gb[['rating']].values).T[0]

from scipy.sparse import csr_matrix
mat_music = csr_matrix((weight,(item, user)))

## Train_test_split

def select(x, samples):
    pair = (x['track_id'],x['user_id'])
    return pair in samples

def make_train(df, ratings, pct_test = 0.2):
    '''
    
    parameters: 
    
    ratings - the original ratings matrix from which you want to generate a train/test set. Test is just a complete
    copy of the original set. This is in the form of a sparse csr_matrix. 
    
    pct_test - The percentage of user-item interactions where an interaction took place that you want to mask in the 
    training set for later comparison to the test set, which contains all of the original ratings. 
    
    returns:
    
    samples - une liste d'intercation à séparer 
    '''
    test_set = ratings.copy() # Make a copy of the original set to be the test set.
    #test_set[test_set != 0] = 1 # Store the test set as a binary preference matrix
    
    training_set = ratings.copy() # Make a copy of the original data we can alter as our training set. 
    nonzero_inds = training_set.nonzero() # Find the indices in the ratings data where an interaction exists
    nonzero_pairs = list(zip(nonzero_inds[0], nonzero_inds[1])) # Zip these pairs together of user,item index into list
    
    random.seed(0) # Set the random seed to zero for reproducibility
    num_samples = int(np.ceil(pct_test*len(nonzero_pairs))) # Round the number of samples needed to the nearest integer
    samples = random.sample(nonzero_pairs, num_samples) # Sample a random number of user-item pairs without replacement
    samples = set(samples)
    
    filtre = df.apply(lambda row: select(row, samples) ,axis = 1)
    
    train_df = df[filtre==False]
    test_df = df[filtre]
    
    return train_df, test_df

train_df, test_df =  make_train(df, mat_music, pct_test=0.3)

### Filtre des track_id
les models mixte colaborativ filtering + content base necessitent un nombre de track_id homogéne entre le train et le test

print('df nb track_id', df['track_id'].nunique())
print('df_train nb track_id', train_df['track_id'].nunique())
print('df_test nb track_id', test_df['track_id'].nunique())

track_to_keep = set(test_df['track_id'].unique())
filter = train_df['track_id'].apply(lambda x: x in track_to_keep)
train_df = train_df[filter]

print('df nb track_id', df['track_id'].nunique())
print('df_train nb track_id', train_df['track_id'].nunique())
print('df_test nb track_id', test_df['track_id'].nunique())

### Verification User_id

print('df nb user_id', df['user_id'].nunique())
print('df_train nb user_id', train_df['user_id'].nunique())
print('df_test nb user_id', test_df['user_id'].nunique())

### Verification Artist_id

print('df nb artist_id', df['artist_id'].nunique())
print('df_train nb artist_id', train_df['artist_id'].nunique())
print('df_test nb artist_id', test_df['artist_id'].nunique())

## Sauvegarde

train_df.to_csv('filtered data/filtered_rating_fm_dataset_train.csv')

test_df.to_csv('filtered data/filtered_rating_fm_dataset_test.csv')

test_df.head()
