==> notebook/3.5 - DATA PREPARATION (user_last_fm).ipynb <==
# DATA PREPARATION (user)

FICHIER: *'userid-profile.tsv'*

## Imports

import pandas as pd
import numpy as np

## Data import

file = 'raw data/lastfm-dataset-1K/userid-profile.tsv'

df =  pd.read_csv(file, sep='\t',error_bad_lines=False, header = None)

# On renome les colonnes
df.rename(columns={0:'user_id', 1:'gender', 2:'age', 3:'country', 4:'registered'}, inplace=True)
df = df.iloc[1:]

df.head()

# On ne garde que la partie numérique des user_id
df['user_id'] = df['user_id'].apply(lambda x: int(x[5:]))

# On regarde l'état des données manquantes
df.isna().sum()

### Regroupemant des pays par régions

df['country'].nunique()

d = {'Europe': ['Bulgaria', 'Slovakia', 'Poland', 'Poland', 'Finland', 'Romania', 'United Kingdom', 'Germany', 'Sweden', 'Italy',
               'Greece','Croatia', 'Serbia', 'Netherlands','Norway', 'Latvia', 'Spain', 'Estonia', 'Czech Republic', 'Portugal', 
                'Switzerland', 'Ireland', 'Belgium', 'France', 'Hungary','Macedonia', 'Slovenia', 'Malta', 'Austria', 'Lithuania',
                'Iceland', 'Northern Mariana Islands', 'Bosnia and Herzegovina'],
     'Asia': ['Japan', "Korea, Democratic People's Republic of", 'Thailand','Singapore', 'China', 'India', ],
     'South America': ['Peru', 'Mexico','Brazil','Chile', 'Venezuela', 'Trinidad and Tobago', 'Nicaragua', 'Argentina',
                       'Netherlands Antilles', 'Colombia'],
     'North America': ['United States', 'Canada', 'United States Minor Outlying Islands', 'Antarctica', ],
     'Russia': ['Russian Federation', 'Belarus'],
     'North Africa': ['Morocco', 'Algeria', 'Tunisia'],
     'Central Africa': ["Cote D'Ivoire", 'Congo, the Democratic Republic of the'],
     'South Africa': ['Zimbabwe'],
     'Meadel Est': ['Turkey', 'Armenia','Israel'],
     'Oceania': ['Australia', 'New Zealand'],
     'Indonesia': ['British Indian Ocean Territory']
     
    
}

def get_region(x):
    
    for key in d:
        if x in d[key]:
            return key
    return np.nan

df['region'] = df['country'].apply(get_region)

## Transformation des dates

# On ne garde que l'année
def get_date(x):
    
    try:
        return x[-5:]
    except:
        return np.nan


df['registered'] = df['registered'].apply(get_date)

df_user = df[['user_id', 'gender','country', 'region', 'registered']]

df_user.head()

## Sauvegarde

df_user.to_csv('filtered data/user_features_fm_dataset.csv')
