==> notebook/3.6 - DATA PREPARATION (features normalisation).ipynb <==
# DATA PREPARATION (Normalisation)

Normalisation des features pour les model content based et mixte (contente + colaboraitf)

## Imports

import numpy as np 
import pandas as pd

## Fichier 'track_features_fm_dataset'

df_track = pd.read_csv('filtered data/track_features_fm_dataset.csv', index_col=0)

### Normalisation et dummies

df_track = pd.concat([df_track,pd.get_dummies(df_track['key'], 'key')], axis=1)
df_track.drop(columns=['key'], inplace=True)

df_track['loudness'] = (df_track['loudness']-df_track['loudness'].min())/(df_track['loudness'].max()-df_track['loudness'].min())

df_track['tempo'] = (df_track['tempo']-df_track['tempo'].min())/(df_track['tempo'].max()-df_track['tempo'].min())

### Verification

df_track.describe()

### Sauvegarde

df_track.to_csv('filtered data/track_features_fm_dataset_normalized.csv')

## Fichier 'user_features_fm_dataset.csv'

df_user= pd.read_csv('filtered data/user_features_fm_dataset.csv', index_col=0)

### Normaliszation et dummies

df_user['gender'] = df_user['gender'].replace({'m':0,'f':1})

df_user = pd.concat([df_user,pd.get_dummies(df_user['country'], 'country')], axis=1)
df_user.drop(columns=['country'], inplace=True)

df_user = pd.concat([df_user,pd.get_dummies(df_user['region'], 'region')], axis=1)
df_user.drop(columns=['region'], inplace=True)

df_user['registered'] = (df_user['registered']-df_user['registered'].min())/(df_user['registered'].max()-df_user['registered'].min())

## Sauvegarde

df_user.to_csv('filtered data/user_features_fm_dataset_normalized.csv')
