==> notebook/4.2 - CONTENT BASED (Fake_users).ipynb <==
# CONTENT BASED RECOMENDATION

Test des model pour la recommendation sur les faux utilisateurs
- User based
- Item based

## Imports

import numpy as np 
import pandas as pd

from sklearn.metrics.pairwise import cosine_similarity
from sklearn.metrics import roc_auc_score

from tqdm import tqdm

import matplotlib.pyplot as plt

## Data Import

df_train = pd.read_csv('filtered data/filtered_rating_fm_dataset_train.csv', index_col=0)

files = ['last_fm_fake_user(1001)_jazz.csv',
         'last_fm_fake_user(1002)_classic.csv',
         'last_fm_fake_user(1003)_pop.csv',
         'last_fm_fake_user(1004)_rock.csv',
         'last_fm_fake_user(1005)_rap.csv']

for file in files:
    df_temp = pd.read_csv(f'filtered data/fake_user/{file}', index_col=0)
    df_temp['rating']=100 # a revoir
    df_train = pd.concat([df_train, df_temp])

df_test = pd.read_csv('filtered data/filtered_rating_fm_dataset_test.csv', index_col=0)

df_track = pd.read_csv('filtered data/track_features_fm_dataset.csv', index_col=0)

## Data preparation

#### On ne garde que les track_id présent dans le df_track

track_ids = set(df_track['track_id'].unique())

filter = df_train['track_id'].apply(lambda x: x in track_ids)
df_train= df_train[filter]

filter = df_test['track_id'].apply(lambda x: x in track_ids)
df_test = df_test[filter]

#### Normalistion et vectorissation des track features

df_track = pd.concat([df_track,pd.get_dummies(df_track['key'], 'key')], axis=1)
df_track.drop(columns=['key'], inplace=True)

df_track['loudness'] = (df_track['loudness']-df_track['loudness'].min())/(df_track['loudness'].max()-df_track['loudness'].min())

df_track['tempo'] = (df_track['tempo']-df_track['tempo'].min())/(df_track['tempo'].max()-df_track['tempo'].min())

df_track['vector'] = df_track.drop(columns=['track_id','artist_name','track_name']).apply(lambda r: tuple(r), axis=1).apply(np.array).values

df_track.head(2)

#### Creation d'un vecteur utilisateur

df_gb = df_train.groupby(['user_id','track_id']).mean().reset_index()
df_gb = df_gb[['user_id','track_id']]

df_user = df_gb.merge(df_track, left_on='track_id', right_on='track_id')
df_user = df_user.drop(columns=['track_id', 'artist_name', 'track_name']).groupby('user_id').agg('mean').reset_index()

df_user['vector'] = df_user.drop(columns=['user_id']).apply(lambda r: tuple(r), axis=1).apply(np.array).values

df_user.head(2)

## User based recomendation

#### Test de la fonction cosine_similarity

# Retourne les n meilleur recommendation en tutilisant la cosine_similarity
def get_reco(vec, n=50):  
        
    track_vec = np.array([x for x in df_track['vector']])
    res = cosine_similarity(vec,track_vec)[0]
    
    df_res = pd.DataFrame()
    df_res['track_id'] = df_track['track_id']
    df_res['cosine_similarity'] = res
    
    if n>=0:
        df_res = df_res.sort_values(by=['cosine_similarity'], ascending=False).iloc[0:n]
    
    return df_res

### Utilisateur Jazz 1001

user = 1001

vect = df_user[df_user['user_id']==user]['vector'].values # On recupére le veteur de notre utilisateur
recomended_tracks = get_reco([vect[0]],10) # On calcule les n=50 meilleurs recomendations

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Calssique 1002

user = 1002

vect = df_user[df_user['user_id']==user]['vector'].values # On recupére le veteur de notre utilisateur
recomended_tracks = get_reco([vect[0]],10) # On calcule les n=50 meilleurs recomendations

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Pop 1003

user = 1003

vect = df_user[df_user['user_id']==user]['vector'].values # On recupére le veteur de notre utilisateur
recomended_tracks = get_reco([vect[0]],10) # On calcule les n=50 meilleurs recomendations

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Rock 1004

user = 1004

vect = df_user[df_user['user_id']==user]['vector'].values # On recupére le veteur de notre utilisateur
recomended_tracks = get_reco([vect[0]],10) # On calcule les n=50 meilleurs recomendations

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Rap 1005

user = 1005

vect = df_user[df_user['user_id']==user]['vector'].values # On recupére le veteur de notre utilisateur
recomended_tracks = get_reco([vect[0]],10) # On calcule les n=50 meilleurs recomendations

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

## Item based recomendation

def get_reco_from_traklist(user_track_vector_list, n=50):
    
    reco_full = pd.DataFrame({'track_id': [], 'cosine_similarity': []})

    for track_vec in user_track_vector_list:
        try:
            if n>=0:
                reco = get_reco([track_vec],50)
            else:
                reco = get_reco([track_vec],-1)
                
            reco_full = pd.concat([reco_full,reco.iloc[1:]])
        except:
            pass
    if n>=0:
        return reco_full.sort_values(by=['cosine_similarity'], ascending=False).iloc[0:n]
    else:
        return reco_full.groupby(['track_id']).mean().reset_index()

def top_user_music(user_id, n=10):
    a = df_train[df_train['user_id']==user_id]
    l = a['track_id']
    return set(l.value_counts()[:n].index)

### Utilisateur Jazz 1001

user = 1001

user_tracks_train = set([x for x in df_train[df_train['user_id']==user]['track_id'].unique()])
user_track_list = pd.DataFrame({'track_id': list(top_user_music(user,10))})
vector_list = pd.merge(user_track_list, df_track, how='left', left_on=['track_id'], right_on = ['track_id'])['vector'].values

recomended_tracks = get_reco_from_traklist(vector_list,10)

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Calssique 1002

user = 1002

user_tracks_train = set([x for x in df_train[df_train['user_id']==user]['track_id'].unique()])
user_track_list = pd.DataFrame({'track_id': list(top_user_music(user,10))})
vector_list = pd.merge(user_track_list, df_track, how='left', left_on=['track_id'], right_on = ['track_id'])['vector'].values

recomended_tracks = get_reco_from_traklist(vector_list,10)

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Pop 1003

user = 1003

user_tracks_train = set([x for x in df_train[df_train['user_id']==user]['track_id'].unique()])
user_track_list = pd.DataFrame({'track_id': list(top_user_music(user,10))})
vector_list = pd.merge(user_track_list, df_track, how='left', left_on=['track_id'], right_on = ['track_id'])['vector'].values

recomended_tracks = get_reco_from_traklist(vector_list,10)

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Rock 1004

user = 1004

user_tracks_train = set([x for x in df_train[df_train['user_id']==user]['track_id'].unique()])
user_track_list = pd.DataFrame({'track_id': list(top_user_music(user,10))})
vector_list = pd.merge(user_track_list, df_track, how='left', left_on=['track_id'], right_on = ['track_id'])['vector'].values

recomended_tracks = get_reco_from_traklist(vector_list,10)

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]

### Utilisateur Rap 1005

user = 1005

user_tracks_train = set([x for x in df_train[df_train['user_id']==user]['track_id'].unique()])
user_track_list = pd.DataFrame({'track_id': list(top_user_music(user,10))})
vector_list = pd.merge(user_track_list, df_track, how='left', left_on=['track_id'], right_on = ['track_id'])['vector'].values

recomended_tracks = get_reco_from_traklist(vector_list,10)

recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['track_id','artist_name','track_name']]
