==> notebook/6.2 - COLLABORATIVE FILTERING (Surprise_SVD Fake Users).ipynb <==
### COLLABORATIVE FILTERING (Surprise SVD)

Test du model SVD sur les faux utilisateurs

### Imports 

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from tqdm import tqdm

from surprise import Dataset
from surprise import Reader

from surprise import SVD

from sklearn.metrics import roc_auc_score

import itertools

## Data imports 

df_train = pd.read_csv('filtered data/filtered_rating_fm_dataset_train.csv', index_col=0)

files = ['last_fm_fake_user(1001)_jazz.csv',
         'last_fm_fake_user(1002)_classic.csv',
         'last_fm_fake_user(1003)_pop.csv',
         'last_fm_fake_user(1004)_rock.csv',
         'last_fm_fake_user(1005)_rap.csv']

for file in files:
    df_temp = pd.read_csv(f'filtered data/fake_user/{file}', index_col=0)
    df_temp['rating']=100 # a revoir
    df_train = pd.concat([df_train, df_temp])

df_train = df_train.groupby(['user_id','track_id']).first().reset_index()

df_test = pd.read_csv('filtered data/filtered_rating_fm_dataset_test.csv', index_col=0)

df_track = df_train.groupby(['track_id']).first().reset_index()

## Data Preparation

reader = Reader(rating_scale=(1, 101))
data = Dataset.load_from_df(df_train[['user_id', 'track_id', 'rating']], reader)

trainset = data.build_full_trainset()

def get_reco(model, df_train, user_id, n=50, filter=False):
    
    scores=[]
    track_ids=[]
    user_ids=[]
    
    user_track = set(df_train[df_train['user_id']==user]['track_id'].unique())

    for track_id in df_train['track_id'].unique():
        score = model.predict(user, track_id, r_ui=1)[3]
        
        if filter and track_id not in user_track:
            user_ids.append(user)
            track_ids.append(track_id)
            scores.append(score)
        elif not filter:
            user_ids.append(user)
            track_ids.append(track_id)
            scores.append(score)
    
    recommendation = pd.DataFrame({'user_id':user_ids,'track_id':track_ids,'score':scores})
    
    if n>=0:
        recommendation = recommendation.sort_values(by=['score'], ascending=False).iloc[:n]
    
    return recommendation
    

## Modèle SVD

svd = SVD(n_factors=150, n_epochs=30)
svd.fit(trainset)

### Utilisateur Jazz 1001

user = 1001

recommendation = get_reco(svd, df_train, user,10)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

user = 1001

recommendation = get_reco(svd, df_train, user,10, filter=True)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

### Utilisateur classique 1002

user = 1002

recommendation = get_reco(svd, df_train, user,10)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

user = 1002

recommendation = get_reco(svd, df_train, user,10, filter=True)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

### Utilisateur Pop 1003

user = 1003

recommendation = get_reco(svd, df_train, user,10)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

user = 1003

recommendation = get_reco(svd, df_train, user,10, filter=True)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

### Utilisateur Rock 1004

user = 1004

recommendation = get_reco(svd, df_train, user,10)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

user = 1004

recommendation = get_reco(svd, df_train, user,10, filter=True)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

### Utilisateur Rap 1005

user = 1005

recommendation = get_reco(svd, df_train, user,10)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]

user = 1005

recommendation = get_reco(svd, df_train, user,10, filter=True)
recomended_tracks = [x for x in recommendation['track_id'].values]

recomended_tracks = pd.DataFrame({'track_id': recomended_tracks})
recomended_tracks.merge(df_track, left_on='track_id', right_on='track_id')[['artist_name','track_name']]
