==> notebook/7.2 - COLLABORATIV FILTERING (Implicit - Fake_users).ipynb <==
# COLLABORATIV FILTERING (implicit Fake_users)

Test du model pour la recommendation sur les faux utilisateurs

## Imports

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from scipy.sparse import coo_matrix

from sklearn.metrics import roc_auc_score

import itertools

from tqdm import tqdm

## Data import

df_train = pd.read_csv('filtered data/filtered_rating_fm_dataset_train.csv', index_col=0)

files = ['last_fm_fake_user(1001)_jazz.csv',
         'last_fm_fake_user(1002)_classic.csv',
         'last_fm_fake_user(1003)_pop.csv',
         'last_fm_fake_user(1004)_rock.csv',
         'last_fm_fake_user(1005)_rap.csv']

for file in files:
    df_temp = pd.read_csv(f'filtered data/fake_user/{file}', index_col=0)
    df_temp['rating']=100 # a revoir
    df_train = pd.concat([df_train, df_temp])

## Data preparation

df_track = df_train.groupby(['track_id']).first().reset_index()

df_gb = df_train.groupby(['user_id','track_id']).mean().reset_index()

item = np.array(df_gb[['track_id']].values).T[0]
user = np.array(df_gb[['user_id']].values).T[0]
weight = np.array(df_gb[['rating']].values).T[0]

mat_music = coo_matrix((weight,(item, user)))
mat_music.shape

## Modèle

import implicit

# Initialisation du modèle
model_music = implicit.als.AlternatingLeastSquares(factors=40,
                                                   use_native=True, 
                                                   use_cg=True, 
                                                   calculate_training_loss=True, 
                                                   num_threads=1, 
                                                   iterations=30)

# Entrainemt du model item/user/confidence weights
model_music.fit(mat_music)

def reco_user_music(user_id, mat, n=10, filter=False):
    if filter:
        f= list(df_train[df_train['user_id']==user_id]['track_id'].unique())
    else:
        f = list()
        
    track_ids = model_music.recommend(user_id, mat.tocsr().T, n, filter_items=f)
    for id,score in track_ids:
        a = df_track[df_track['track_id']==id]
        track = a['track_name'].values[0]
        artist = a['artist_name'].values[0]
        print(f'{str(score)[:4]}--> {artist}: {track}')

### Utilisateur Jazz 1001

reco_user_music(1001, mat_music, filter=False)

reco_user_music(1001, mat_music, filter=True)

### Utilisateur Classique 1002

reco_user_music(1002, mat_music, filter=False)

reco_user_music(1002, mat_music, filter=True)

### Utilisateur Pop 1003

reco_user_music(1003, mat_music, filter=False)

reco_user_music(1003, mat_music, filter=True)

### Utilisateur Rock 1004

reco_user_music(1004, mat_music, filter=False)

reco_user_music(1004, mat_music, filter=True)

### Utilisateur Rap 1005

reco_user_music(1005, mat_music, filter=False)

reco_user_music(1005, mat_music, filter=True)
